+++ src/noahtheduke/splint.clj
@@ -19,6 +19,7 @@
 
    noahtheduke.splint.rules.lint.assoc-fn
    noahtheduke.splint.rules.lint.body-unquote-splicing
+   noahtheduke.splint.rules.lint.defmethod-names
    noahtheduke.splint.rules.lint.divide-by-one
    noahtheduke.splint.rules.lint.dorun-map
    noahtheduke.splint.rules.lint.dot-class-method